fix: userId추가해서 에러핸들링 추가

+++ api/services/goods.service.js
@@ -1,11 +1,22 @@
 const GoodsRepository = require("../repositories/goods.repository");
+// const LikesRepository = require("../repositories/")
 const { makeError } = require("../error");
 
 class GoodsService {
   goodsRepository = new GoodsRepository();
+  // likesRepository = new LikesRepository()
 
-  createGoods = async (name, content, price, category, option, ...file) => {
+  createGoods = async (
+    userId,
+    name,
+    content,
+    price,
+    category,
+    option,
+    ...file
+  ) => {
     const createGoodsData = await this.goodsRepository.createGoods(
+      userId,
       name,
       content,
       price,
@@ -33,8 +44,9 @@ class GoodsService {
     return goodsData;
   };
 
-  getGoods = async (goodsId) => {
-    const goodsData = await this.goodsRepository.getGoods(goodsId);
+  getGoods = async (userId, goodsId) => {
+    const goodsData = await this.goodsRepository.getGoods(userId, goodsId);
+    // const likeStatus = await this.goodsRepository
     if (!goodsData) {
       throw new makeError({
         message: "게시글 조회를 실패했습니다.",
@@ -44,7 +56,15 @@ class GoodsService {
     return goodsData;
   };
 
-  editGoods = async (goodsId, name, content) => {
+  editGoods = async (
+    userId,
+    goodsId,
+    name,
+    content,
+    price,
+    category,
+    option
+  ) => {
     const goodsData = await this.goodsRepository.getGoods(goodsId);
     if (!goodsData) {
       throw new makeError({
@@ -52,16 +72,35 @@ class GoodsService {
         code: 404,
       });
     }
-    await this.goodsRepository.editGoods(goodsId, name, content);
+    if (goodsData.userId !== userId) {
+      throw new makeError({
+        message: "게시글 수정 권한이 없습니다.",
+        code: 401,
+      });
+    }
+    await this.goodsRepository.editGoods(
+      goodsId,
+      name,
+      content,
+      price,
+      category,
+      option
+    );
     return;
   };
 
-  deleteGoods = async (goodsId) => {
+  deleteGoods = async (userId, goodsId) => {
     const goodsData = await this.goodsRepository.getGoods(goodsId);
     if (!goodsData) {
       throw new makeError({
         message: "삭제할 게시글을 찾지 못했습니다.",
         code: 404,
+      });
+    }
+    if (goodsData.userId !== userId) {
+      throw new makeError({
+        message: "이 게시물을 삭제할 권한이 없습니다.",
+        code: 401,
       });
     }
     await this.goodsRepository.deleteGoods(goodsId);
